Struct aws_sdk_iam::client::fluent_builders::ListInstanceProfilesForRole

pub struct ListInstanceProfilesForRole<C = DynConnector> { /* fields omitted */ }

Implementations

impl<C> ListInstanceProfilesForRole<C>

pub async fn send(
self
) -> Result<ListInstanceProfilesForRoleOutput, SdkError<ListInstanceProfilesForRoleError>> where
C: SmithyConnector,

pub fn role_name(self, input: impl Into<String>) -> Self

The name of the role to list instance profiles for.

This parameter allows (through its regex pattern) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: _+=,.@-

pub fn set_role_name(self, input: Option<String>) -> Self

pub fn marker(self, input: impl Into<String>) -> Self

Use this parameter only when paginating results and only after you receive a response indicating that the results are truncated. Set it to the value of the Marker element in the response that you received to indicate where the next call should start.

pub fn set_marker(self, input: Option<String>) -> Self

pub fn max_items(self, input: i32) -> Self

Use this only when paginating results to indicate the maximum number of items you want in the response. If additional items exist beyond the maximum you specify, the IsTruncated response element is true.

If you do not include this parameter, the number of items defaults to 100. Note that IAM might return fewer results, even when there are more results available. In that case, the IsTruncated response element returns true, and Marker contains a value to include in the subsequent call that tells the service where to continue from.

pub fn set_max_items(self, input: Option<i32>) -> Self