Struct aws_sdk_iam::client::fluent_builders::UpdateAccountPasswordPolicy

pub struct UpdateAccountPasswordPolicy<C = DynConnector> { /* fields omitted */ }

Implementations

impl<C> UpdateAccountPasswordPolicy<C>

pub async fn send(
self
) -> Result<UpdateAccountPasswordPolicyOutput, SdkError<UpdateAccountPasswordPolicyError>> where
C: SmithyConnector,

pub fn minimum_password_length(self, input: i32) -> Self

The minimum number of characters allowed in an IAM user password.

If you do not specify a value for this parameter, then the operation uses the default value of 6.

pub fn set_minimum_password_length(self, input: Option<i32>) -> Self

pub fn require_symbols(self, input: bool) -> Self

Specifies whether IAM user passwords must contain at least one of the following non-alphanumeric characters:

! @ # $ % ^ & * ( ) _ + - = [ ] { } | '

If you do not specify a value for this parameter, then the operation uses the default value of false. The result is that passwords do not require at least one symbol character.

pub fn set_require_symbols(self, input: Option<bool>) -> Self

pub fn require_numbers(self, input: bool) -> Self

Specifies whether IAM user passwords must contain at least one numeric character (0 to 9).

If you do not specify a value for this parameter, then the operation uses the default value of false. The result is that passwords do not require at least one numeric character.

pub fn set_require_numbers(self, input: Option<bool>) -> Self

pub fn require_uppercase_characters(self, input: bool) -> Self

Specifies whether IAM user passwords must contain at least one uppercase character from the ISO basic Latin alphabet (A to Z).

If you do not specify a value for this parameter, then the operation uses the default value of false. The result is that passwords do not require at least one uppercase character.

pub fn set_require_uppercase_characters(self, input: Option<bool>) -> Self

pub fn require_lowercase_characters(self, input: bool) -> Self

Specifies whether IAM user passwords must contain at least one lowercase character from the ISO basic Latin alphabet (a to z).

If you do not specify a value for this parameter, then the operation uses the default value of false. The result is that passwords do not require at least one lowercase character.

pub fn set_require_lowercase_characters(self, input: Option<bool>) -> Self

pub fn allow_users_to_change_password(self, input: bool) -> Self

Allows all IAM users in your account to use the AWS Management Console to change their own passwords. For more information, see Letting IAM users change their own passwords in the IAM User Guide.

If you do not specify a value for this parameter, then the operation uses the default value of false. The result is that IAM users in the account do not automatically have permissions to change their own password.

pub fn set_allow_users_to_change_password(self, input: Option<bool>) -> Self

pub fn max_password_age(self, input: i32) -> Self

The number of days that an IAM user password is valid.

If you do not specify a value for this parameter, then the operation uses the default value of 0. The result is that IAM user passwords never expire.

pub fn set_max_password_age(self, input: Option<i32>) -> Self

pub fn password_reuse_prevention(self, input: i32) -> Self

Specifies the number of previous passwords that IAM users are prevented from reusing.

If you do not specify a value for this parameter, then the operation uses the default value of 0. The result is that IAM users are not prevented from reusing previous passwords.

pub fn set_password_reuse_prevention(self, input: Option<i32>) -> Self

pub fn hard_expiry(self, input: bool) -> Self

Prevents IAM users from setting a new password after their password has expired. The IAM user cannot be accessed until an administrator resets the password.

If you do not specify a value for this parameter, then the operation uses the default value of false. The result is that IAM users can change their passwords after they expire and continue to sign in as the user.

pub fn set_hard_expiry(self, input: Option<bool>) -> Self