Struct aws_sdk_iam::client::Client

pub struct Client<C = DynConnector, M = AwsMiddleware, R = Standard> { /* fields omitted */ }

An ergonomic service client for AWSIdentityManagementV20100508.

This client allows ergonomic access to a AWSIdentityManagementV20100508-shaped service. Each method corresponds to an endpoint defined in the service’s Smithy model, and the request and response shapes are auto-generated from that same model.

Using a Client

Once you have a client set up, you can access the service’s endpoints by calling the appropriate method on Client. Each such method returns a request builder for that endpoint, with methods for setting the various fields of the request. Once your request is complete, use the send method to send the request. send returns a future, which you then have to .await to get the service’s response.

Implementations

impl<C, M, R> Client<C, M, R>

pub fn with_config(client: Client<C, M, R>, conf: Config) -> Self

pub fn conf(&self) -> &Config

impl<C, M, R> Client<C, M, R> where
C: SmithyConnector,
M: SmithyMiddleware<C>,
R: NewRequestPolicy,

pub fn add_client_id_to_open_id_connect_provider(
&self
) -> AddClientIDToOpenIDConnectProvider<C, M, R>

pub fn add_role_to_instance_profile(&self) -> AddRoleToInstanceProfile<C, M, R>

pub fn add_user_to_group(&self) -> AddUserToGroup<C, M, R>

pub fn attach_group_policy(&self) -> AttachGroupPolicy<C, M, R>

pub fn attach_role_policy(&self) -> AttachRolePolicy<C, M, R>

pub fn attach_user_policy(&self) -> AttachUserPolicy<C, M, R>

pub fn change_password(&self) -> ChangePassword<C, M, R>

pub fn create_access_key(&self) -> CreateAccessKey<C, M, R>

pub fn create_account_alias(&self) -> CreateAccountAlias<C, M, R>

pub fn create_group(&self) -> CreateGroup<C, M, R>

pub fn create_instance_profile(&self) -> CreateInstanceProfile<C, M, R>

pub fn create_login_profile(&self) -> CreateLoginProfile<C, M, R>

pub fn create_open_id_connect_provider(
&self
) -> CreateOpenIDConnectProvider<C, M, R>

pub fn create_policy(&self) -> CreatePolicy<C, M, R>

pub fn create_policy_version(&self) -> CreatePolicyVersion<C, M, R>

pub fn create_role(&self) -> CreateRole<C, M, R>

pub fn create_saml_provider(&self) -> CreateSAMLProvider<C, M, R>

pub fn create_service_linked_role(&self) -> CreateServiceLinkedRole<C, M, R>

pub fn create_service_specific_credential(
&self
) -> CreateServiceSpecificCredential<C, M, R>

pub fn create_user(&self) -> CreateUser<C, M, R>

pub fn create_virtual_mfa_device(&self) -> CreateVirtualMFADevice<C, M, R>

pub fn deactivate_mfa_device(&self) -> DeactivateMFADevice<C, M, R>

pub fn delete_access_key(&self) -> DeleteAccessKey<C, M, R>

pub fn delete_account_alias(&self) -> DeleteAccountAlias<C, M, R>

pub fn delete_account_password_policy(
&self
) -> DeleteAccountPasswordPolicy<C, M, R>

pub fn delete_group(&self) -> DeleteGroup<C, M, R>

pub fn delete_group_policy(&self) -> DeleteGroupPolicy<C, M, R>

pub fn delete_instance_profile(&self) -> DeleteInstanceProfile<C, M, R>

pub fn delete_login_profile(&self) -> DeleteLoginProfile<C, M, R>

pub fn delete_open_id_connect_provider(
&self
) -> DeleteOpenIDConnectProvider<C, M, R>

pub fn delete_policy(&self) -> DeletePolicy<C, M, R>

pub fn delete_policy_version(&self) -> DeletePolicyVersion<C, M, R>

pub fn delete_role(&self) -> DeleteRole<C, M, R>

pub fn delete_role_permissions_boundary(
&self
) -> DeleteRolePermissionsBoundary<C, M, R>

pub fn delete_role_policy(&self) -> DeleteRolePolicy<C, M, R>

pub fn delete_saml_provider(&self) -> DeleteSAMLProvider<C, M, R>

pub fn delete_server_certificate(&self) -> DeleteServerCertificate<C, M, R>

pub fn delete_service_linked_role(&self) -> DeleteServiceLinkedRole<C, M, R>

pub fn delete_service_specific_credential(
&self
) -> DeleteServiceSpecificCredential<C, M, R>

pub fn delete_signing_certificate(&self) -> DeleteSigningCertificate<C, M, R>

pub fn delete_ssh_public_key(&self) -> DeleteSSHPublicKey<C, M, R>

pub fn delete_user(&self) -> DeleteUser<C, M, R>

pub fn delete_user_permissions_boundary(
&self
) -> DeleteUserPermissionsBoundary<C, M, R>

pub fn delete_user_policy(&self) -> DeleteUserPolicy<C, M, R>

pub fn delete_virtual_mfa_device(&self) -> DeleteVirtualMFADevice<C, M, R>

pub fn detach_group_policy(&self) -> DetachGroupPolicy<C, M, R>

pub fn detach_role_policy(&self) -> DetachRolePolicy<C, M, R>

pub fn detach_user_policy(&self) -> DetachUserPolicy<C, M, R>

pub fn enable_mfa_device(&self) -> EnableMFADevice<C, M, R>

pub fn generate_credential_report(&self) -> GenerateCredentialReport<C, M, R>

pub fn generate_organizations_access_report(
&self
) -> GenerateOrganizationsAccessReport<C, M, R>

pub fn generate_service_last_accessed_details(
&self
) -> GenerateServiceLastAccessedDetails<C, M, R>

pub fn get_access_key_last_used(&self) -> GetAccessKeyLastUsed<C, M, R>

pub fn get_account_authorization_details(
&self
) -> GetAccountAuthorizationDetails<C, M, R>

pub fn get_account_password_policy(&self) -> GetAccountPasswordPolicy<C, M, R>

pub fn get_account_summary(&self) -> GetAccountSummary<C, M, R>

pub fn get_context_keys_for_custom_policy(
&self
) -> GetContextKeysForCustomPolicy<C, M, R>

pub fn get_context_keys_for_principal_policy(
&self
) -> GetContextKeysForPrincipalPolicy<C, M, R>

pub fn get_credential_report(&self) -> GetCredentialReport<C, M, R>

pub fn get_group(&self) -> GetGroup<C, M, R>

pub fn get_group_policy(&self) -> GetGroupPolicy<C, M, R>

pub fn get_instance_profile(&self) -> GetInstanceProfile<C, M, R>

pub fn get_login_profile(&self) -> GetLoginProfile<C, M, R>

pub fn get_open_id_connect_provider(&self) -> GetOpenIDConnectProvider<C, M, R>

pub fn get_organizations_access_report(
&self
) -> GetOrganizationsAccessReport<C, M, R>

pub fn get_policy(&self) -> GetPolicy<C, M, R>

pub fn get_policy_version(&self) -> GetPolicyVersion<C, M, R>

pub fn get_role(&self) -> GetRole<C, M, R>

pub fn get_role_policy(&self) -> GetRolePolicy<C, M, R>

pub fn get_saml_provider(&self) -> GetSAMLProvider<C, M, R>

pub fn get_server_certificate(&self) -> GetServerCertificate<C, M, R>

pub fn get_service_last_accessed_details(
&self
) -> GetServiceLastAccessedDetails<C, M, R>

pub fn get_service_last_accessed_details_with_entities(
&self
) -> GetServiceLastAccessedDetailsWithEntities<C, M, R>

pub fn get_service_linked_role_deletion_status(
&self
) -> GetServiceLinkedRoleDeletionStatus<C, M, R>

pub fn get_ssh_public_key(&self) -> GetSSHPublicKey<C, M, R>

pub fn get_user(&self) -> GetUser<C, M, R>

pub fn get_user_policy(&self) -> GetUserPolicy<C, M, R>

pub fn list_access_keys(&self) -> ListAccessKeys<C, M, R>

pub fn list_account_aliases(&self) -> ListAccountAliases<C, M, R>

pub fn list_attached_group_policies(&self) -> ListAttachedGroupPolicies<C, M, R>

pub fn list_attached_role_policies(&self) -> ListAttachedRolePolicies<C, M, R>

pub fn list_attached_user_policies(&self) -> ListAttachedUserPolicies<C, M, R>

pub fn list_entities_for_policy(&self) -> ListEntitiesForPolicy<C, M, R>

pub fn list_group_policies(&self) -> ListGroupPolicies<C, M, R>

pub fn list_groups(&self) -> ListGroups<C, M, R>

pub fn list_groups_for_user(&self) -> ListGroupsForUser<C, M, R>

pub fn list_instance_profiles(&self) -> ListInstanceProfiles<C, M, R>

pub fn list_instance_profiles_for_role(
&self
) -> ListInstanceProfilesForRole<C, M, R>

pub fn list_instance_profile_tags(&self) -> ListInstanceProfileTags<C, M, R>

pub fn list_mfa_devices(&self) -> ListMFADevices<C, M, R>

pub fn list_mfa_device_tags(&self) -> ListMFADeviceTags<C, M, R>

pub fn list_open_id_connect_providers(
&self
) -> ListOpenIDConnectProviders<C, M, R>

pub fn list_open_id_connect_provider_tags(
&self
) -> ListOpenIDConnectProviderTags<C, M, R>

pub fn list_policies(&self) -> ListPolicies<C, M, R>

pub fn list_policies_granting_service_access(
&self
) -> ListPoliciesGrantingServiceAccess<C, M, R>

pub fn list_policy_tags(&self) -> ListPolicyTags<C, M, R>

pub fn list_policy_versions(&self) -> ListPolicyVersions<C, M, R>

pub fn list_role_policies(&self) -> ListRolePolicies<C, M, R>

pub fn list_roles(&self) -> ListRoles<C, M, R>

pub fn list_role_tags(&self) -> ListRoleTags<C, M, R>

pub fn list_saml_providers(&self) -> ListSAMLProviders<C, M, R>

pub fn list_saml_provider_tags(&self) -> ListSAMLProviderTags<C, M, R>

pub fn list_server_certificates(&self) -> ListServerCertificates<C, M, R>

pub fn list_server_certificate_tags(&self) -> ListServerCertificateTags<C, M, R>

pub fn list_service_specific_credentials(
&self
) -> ListServiceSpecificCredentials<C, M, R>

pub fn list_signing_certificates(&self) -> ListSigningCertificates<C, M, R>

pub fn list_ssh_public_keys(&self) -> ListSSHPublicKeys<C, M, R>

pub fn list_user_policies(&self) -> ListUserPolicies<C, M, R>

pub fn list_users(&self) -> ListUsers<C, M, R>

pub fn list_user_tags(&self) -> ListUserTags<C, M, R>

pub fn list_virtual_mfa_devices(&self) -> ListVirtualMFADevices<C, M, R>

pub fn put_group_policy(&self) -> PutGroupPolicy<C, M, R>

pub fn put_role_permissions_boundary(
&self
) -> PutRolePermissionsBoundary<C, M, R>

pub fn put_role_policy(&self) -> PutRolePolicy<C, M, R>

pub fn put_user_permissions_boundary(
&self
) -> PutUserPermissionsBoundary<C, M, R>

pub fn put_user_policy(&self) -> PutUserPolicy<C, M, R>

pub fn remove_client_id_from_open_id_connect_provider(
&self
) -> RemoveClientIDFromOpenIDConnectProvider<C, M, R>

pub fn remove_role_from_instance_profile(
&self
) -> RemoveRoleFromInstanceProfile<C, M, R>

pub fn remove_user_from_group(&self) -> RemoveUserFromGroup<C, M, R>

pub fn reset_service_specific_credential(
&self
) -> ResetServiceSpecificCredential<C, M, R>

pub fn resync_mfa_device(&self) -> ResyncMFADevice<C, M, R>

pub fn set_default_policy_version(&self) -> SetDefaultPolicyVersion<C, M, R>

pub fn set_security_token_service_preferences(
&self
) -> SetSecurityTokenServicePreferences<C, M, R>

pub fn simulate_custom_policy(&self) -> SimulateCustomPolicy<C, M, R>

pub fn simulate_principal_policy(&self) -> SimulatePrincipalPolicy<C, M, R>

pub fn tag_instance_profile(&self) -> TagInstanceProfile<C, M, R>

pub fn tag_mfa_device(&self) -> TagMFADevice<C, M, R>

pub fn tag_open_id_connect_provider(&self) -> TagOpenIDConnectProvider<C, M, R>

pub fn tag_policy(&self) -> TagPolicy<C, M, R>

pub fn tag_role(&self) -> TagRole<C, M, R>

pub fn tag_saml_provider(&self) -> TagSAMLProvider<C, M, R>

pub fn tag_server_certificate(&self) -> TagServerCertificate<C, M, R>

pub fn tag_user(&self) -> TagUser<C, M, R>

pub fn untag_instance_profile(&self) -> UntagInstanceProfile<C, M, R>

pub fn untag_mfa_device(&self) -> UntagMFADevice<C, M, R>

pub fn untag_open_id_connect_provider(
&self
) -> UntagOpenIDConnectProvider<C, M, R>

pub fn untag_policy(&self) -> UntagPolicy<C, M, R>

pub fn untag_role(&self) -> UntagRole<C, M, R>

pub fn untag_saml_provider(&self) -> UntagSAMLProvider<C, M, R>

pub fn untag_server_certificate(&self) -> UntagServerCertificate<C, M, R>

pub fn untag_user(&self) -> UntagUser<C, M, R>

pub fn update_access_key(&self) -> UpdateAccessKey<C, M, R>

pub fn update_account_password_policy(
&self
) -> UpdateAccountPasswordPolicy<C, M, R>

pub fn update_assume_role_policy(&self) -> UpdateAssumeRolePolicy<C, M, R>

pub fn update_group(&self) -> UpdateGroup<C, M, R>

pub fn update_login_profile(&self) -> UpdateLoginProfile<C, M, R>

pub fn update_open_id_connect_provider_thumbprint(
&self
) -> UpdateOpenIDConnectProviderThumbprint<C, M, R>

pub fn update_role(&self) -> UpdateRole<C, M, R>

pub fn update_role_description(&self) -> UpdateRoleDescription<C, M, R>

pub fn update_saml_provider(&self) -> UpdateSAMLProvider<C, M, R>

pub fn update_server_certificate(&self) -> UpdateServerCertificate<C, M, R>

pub fn update_service_specific_credential(
&self
) -> UpdateServiceSpecificCredential<C, M, R>

pub fn update_signing_certificate(&self) -> UpdateSigningCertificate<C, M, R>

pub fn update_ssh_public_key(&self) -> UpdateSSHPublicKey<C, M, R>

pub fn update_user(&self) -> UpdateUser<C, M, R>

pub fn upload_server_certificate(&self) -> UploadServerCertificate<C, M, R>

pub fn upload_signing_certificate(&self) -> UploadSigningCertificate<C, M, R>

pub fn upload_ssh_public_key(&self) -> UploadSSHPublicKey<C, M, R>

impl<C> Client<C, AwsMiddleware, Standard>

pub fn from_conf_conn(conf: Config, conn: C) -> Self

impl Client<DynConnector, AwsMiddleware, Standard>

pub fn new(config: &Config) -> Self

pub fn from_conf(conf: Config) -> Self