Module aws_sdk_iam::error

Modules

concurrent_modification_exception

See ConcurrentModificationException

credential_report_expired_exception

See CredentialReportExpiredException

credential_report_not_present_exception

See CredentialReportNotPresentException

credential_report_not_ready_exception

See CredentialReportNotReadyException

delete_conflict_exception

See DeleteConflictException

duplicate_certificate_exception

See DuplicateCertificateException

duplicate_ssh_public_key_exception

See DuplicateSshPublicKeyException

entity_already_exists_exception

See EntityAlreadyExistsException

entity_temporarily_unmodifiable_exception

See EntityTemporarilyUnmodifiableException

invalid_authentication_code_exception

See InvalidAuthenticationCodeException

invalid_certificate_exception

See InvalidCertificateException

invalid_input_exception

See InvalidInputException

invalid_public_key_exception

See InvalidPublicKeyException

invalid_user_type_exception

See InvalidUserTypeException

key_pair_mismatch_exception

See KeyPairMismatchException

limit_exceeded_exception

See LimitExceededException

malformed_certificate_exception

See MalformedCertificateException

malformed_policy_document_exception

See MalformedPolicyDocumentException

no_such_entity_exception

See NoSuchEntityException

password_policy_violation_exception

See PasswordPolicyViolationException

policy_evaluation_exception

See PolicyEvaluationException

policy_not_attachable_exception

See PolicyNotAttachableException

report_generation_limit_exceeded_exception

See ReportGenerationLimitExceededException

service_failure_exception

See ServiceFailureException

service_not_supported_exception

See ServiceNotSupportedException

unmodifiable_entity_exception

See UnmodifiableEntityException

unrecognized_public_key_encoding_exception

See UnrecognizedPublicKeyEncodingException

Structs

AddClientIDToOpenIDConnectProviderError

AddRoleToInstanceProfileError

AddUserToGroupError

AttachGroupPolicyError

AttachRolePolicyError

AttachUserPolicyError

ChangePasswordError

ConcurrentModificationException

The request was rejected because multiple requests to change this object were submitted

CreateAccessKeyError

CreateAccountAliasError

CreateGroupError

CreateInstanceProfileError

CreateLoginProfileError

CreateOpenIDConnectProviderError

CreatePolicyError

CreatePolicyVersionError

CreateRoleError

CreateSAMLProviderError

CreateServiceLinkedRoleError

CreateServiceSpecificCredentialError

CreateUserError

CreateVirtualMFADeviceError

CredentialReportExpiredException

The request was rejected because the most recent credential report has expired. To

CredentialReportNotPresentException

The request was rejected because the credential report does not exist. To generate a

CredentialReportNotReadyException

The request was rejected because the credential report is still being generated.

DeactivateMFADeviceError

DeleteAccessKeyError

DeleteAccountAliasError

DeleteAccountPasswordPolicyError

DeleteConflictException

The request was rejected because it attempted to delete a resource that has attached

DeleteGroupError

DeleteGroupPolicyError

DeleteInstanceProfileError

DeleteLoginProfileError

DeleteOpenIDConnectProviderError

DeletePolicyError

DeletePolicyVersionError

DeleteRoleError

DeleteRolePermissionsBoundaryError

DeleteRolePolicyError

DeleteSAMLProviderError

DeleteSSHPublicKeyError

DeleteServerCertificateError

DeleteServiceLinkedRoleError

DeleteServiceSpecificCredentialError

DeleteSigningCertificateError

DeleteUserError

DeleteUserPermissionsBoundaryError

DeleteUserPolicyError

DeleteVirtualMFADeviceError

DetachGroupPolicyError

DetachRolePolicyError

DetachUserPolicyError

DuplicateCertificateException

The request was rejected because the same certificate is associated with an IAM user in

DuplicateSshPublicKeyException

The request was rejected because the SSH public key is already associated with the

EnableMFADeviceError

EntityAlreadyExistsException

The request was rejected because it attempted to create a resource that already

EntityTemporarilyUnmodifiableException

The request was rejected because it referenced an entity that is temporarily unmodifiable,

GenerateCredentialReportError

GenerateOrganizationsAccessReportError

GenerateServiceLastAccessedDetailsError

GetAccessKeyLastUsedError

GetAccountAuthorizationDetailsError

GetAccountPasswordPolicyError

GetAccountSummaryError

GetContextKeysForCustomPolicyError

GetContextKeysForPrincipalPolicyError

GetCredentialReportError

GetGroupError

GetGroupPolicyError

GetInstanceProfileError

GetLoginProfileError

GetOpenIDConnectProviderError

GetOrganizationsAccessReportError

GetPolicyError

GetPolicyVersionError

GetRoleError

GetRolePolicyError

GetSAMLProviderError

GetSSHPublicKeyError

GetServerCertificateError

GetServiceLastAccessedDetailsError

GetServiceLastAccessedDetailsWithEntitiesError

GetServiceLinkedRoleDeletionStatusError

GetUserError

GetUserPolicyError

InvalidAuthenticationCodeException

The request was rejected because the authentication code was not recognized. The error

InvalidCertificateException

The request was rejected because the certificate is invalid.

InvalidInputException

The request was rejected because an invalid or out-of-range value was supplied for an

InvalidPublicKeyException

The request was rejected because the public key is malformed or otherwise invalid.

InvalidUserTypeException

The request was rejected because the type of user for the transaction was

KeyPairMismatchException

The request was rejected because the public key certificate and the private key do not

LimitExceededException

The request was rejected because it attempted to create resources beyond the current AWS

ListAccessKeysError

ListAccountAliasesError

ListAttachedGroupPoliciesError

ListAttachedRolePoliciesError

ListAttachedUserPoliciesError

ListEntitiesForPolicyError

ListGroupPoliciesError

ListGroupsError

ListGroupsForUserError

ListInstanceProfileTagsError

ListInstanceProfilesError

ListInstanceProfilesForRoleError

ListMFADeviceTagsError

ListMFADevicesError

ListOpenIDConnectProviderTagsError

ListOpenIDConnectProvidersError

ListPoliciesError

ListPoliciesGrantingServiceAccessError

ListPolicyTagsError

ListPolicyVersionsError

ListRolePoliciesError

ListRoleTagsError

ListRolesError

ListSAMLProviderTagsError

ListSAMLProvidersError

ListSSHPublicKeysError

ListServerCertificateTagsError

ListServerCertificatesError

ListServiceSpecificCredentialsError

ListSigningCertificatesError

ListUserPoliciesError

ListUserTagsError

ListUsersError

ListVirtualMFADevicesError

MalformedCertificateException

The request was rejected because the certificate was malformed or expired. The error

MalformedPolicyDocumentException

The request was rejected because the policy document was malformed. The error message

NoSuchEntityException

The request was rejected because it referenced a resource entity that does not exist. The

PasswordPolicyViolationException

The request was rejected because the provided password did not meet the requirements

PolicyEvaluationException

The request failed because a provided policy could not be successfully evaluated. An

PolicyNotAttachableException

The request failed because AWS service role policies can only be attached to the

PutGroupPolicyError

PutRolePermissionsBoundaryError

PutRolePolicyError

PutUserPermissionsBoundaryError

PutUserPolicyError

RemoveClientIDFromOpenIDConnectProviderError

RemoveRoleFromInstanceProfileError

RemoveUserFromGroupError

ReportGenerationLimitExceededException

The request failed because the maximum number of concurrent requests for this account are

ResetServiceSpecificCredentialError

ResyncMFADeviceError

ServiceFailureException

The request processing has failed because of an unknown error, exception or

ServiceNotSupportedException

The specified service does not support service-specific credentials.

SetDefaultPolicyVersionError

SetSecurityTokenServicePreferencesError

SimulateCustomPolicyError

SimulatePrincipalPolicyError

TagInstanceProfileError

TagMFADeviceError

TagOpenIDConnectProviderError

TagPolicyError

TagRoleError

TagSAMLProviderError

TagServerCertificateError

TagUserError

UnmodifiableEntityException

The request was rejected because only the service that depends on the service-linked role

UnrecognizedPublicKeyEncodingException

The request was rejected because the public key encoding format is unsupported or

UntagInstanceProfileError

UntagMFADeviceError

UntagOpenIDConnectProviderError

UntagPolicyError

UntagRoleError

UntagSAMLProviderError

UntagServerCertificateError

UntagUserError

UpdateAccessKeyError

UpdateAccountPasswordPolicyError

UpdateAssumeRolePolicyError

UpdateGroupError

UpdateLoginProfileError

UpdateOpenIDConnectProviderThumbprintError

UpdateRoleDescriptionError

UpdateRoleError

UpdateSAMLProviderError

UpdateSSHPublicKeyError

UpdateServerCertificateError

UpdateServiceSpecificCredentialError

UpdateSigningCertificateError

UpdateUserError

UploadSSHPublicKeyError

UploadServerCertificateError

UploadSigningCertificateError

Enums

AddClientIDToOpenIDConnectProviderErrorKind

AddRoleToInstanceProfileErrorKind

AddUserToGroupErrorKind

AttachGroupPolicyErrorKind

AttachRolePolicyErrorKind

AttachUserPolicyErrorKind

ChangePasswordErrorKind

CreateAccessKeyErrorKind

CreateAccountAliasErrorKind

CreateGroupErrorKind

CreateInstanceProfileErrorKind

CreateLoginProfileErrorKind

CreateOpenIDConnectProviderErrorKind

CreatePolicyErrorKind

CreatePolicyVersionErrorKind

CreateRoleErrorKind

CreateSAMLProviderErrorKind

CreateServiceLinkedRoleErrorKind

CreateServiceSpecificCredentialErrorKind

CreateUserErrorKind

CreateVirtualMFADeviceErrorKind

DeactivateMFADeviceErrorKind

DeleteAccessKeyErrorKind

DeleteAccountAliasErrorKind

DeleteAccountPasswordPolicyErrorKind

DeleteGroupErrorKind

DeleteGroupPolicyErrorKind

DeleteInstanceProfileErrorKind

DeleteLoginProfileErrorKind

DeleteOpenIDConnectProviderErrorKind

DeletePolicyErrorKind

DeletePolicyVersionErrorKind

DeleteRoleErrorKind

DeleteRolePermissionsBoundaryErrorKind

DeleteRolePolicyErrorKind

DeleteSAMLProviderErrorKind

DeleteSSHPublicKeyErrorKind

DeleteServerCertificateErrorKind

DeleteServiceLinkedRoleErrorKind

DeleteServiceSpecificCredentialErrorKind

DeleteSigningCertificateErrorKind

DeleteUserErrorKind

DeleteUserPermissionsBoundaryErrorKind

DeleteUserPolicyErrorKind

DeleteVirtualMFADeviceErrorKind

DetachGroupPolicyErrorKind

DetachRolePolicyErrorKind

DetachUserPolicyErrorKind

EnableMFADeviceErrorKind

GenerateCredentialReportErrorKind

GenerateOrganizationsAccessReportErrorKind

GenerateServiceLastAccessedDetailsErrorKind

GetAccessKeyLastUsedErrorKind

GetAccountAuthorizationDetailsErrorKind

GetAccountPasswordPolicyErrorKind

GetAccountSummaryErrorKind

GetContextKeysForCustomPolicyErrorKind

GetContextKeysForPrincipalPolicyErrorKind

GetCredentialReportErrorKind

GetGroupErrorKind

GetGroupPolicyErrorKind

GetInstanceProfileErrorKind

GetLoginProfileErrorKind

GetOpenIDConnectProviderErrorKind

GetOrganizationsAccessReportErrorKind

GetPolicyErrorKind

GetPolicyVersionErrorKind

GetRoleErrorKind

GetRolePolicyErrorKind

GetSAMLProviderErrorKind

GetSSHPublicKeyErrorKind

GetServerCertificateErrorKind

GetServiceLastAccessedDetailsErrorKind

GetServiceLastAccessedDetailsWithEntitiesErrorKind

GetServiceLinkedRoleDeletionStatusErrorKind

GetUserErrorKind

GetUserPolicyErrorKind

ListAccessKeysErrorKind

ListAccountAliasesErrorKind

ListAttachedGroupPoliciesErrorKind

ListAttachedRolePoliciesErrorKind

ListAttachedUserPoliciesErrorKind

ListEntitiesForPolicyErrorKind

ListGroupPoliciesErrorKind

ListGroupsErrorKind

ListGroupsForUserErrorKind

ListInstanceProfileTagsErrorKind

ListInstanceProfilesErrorKind

ListInstanceProfilesForRoleErrorKind

ListMFADeviceTagsErrorKind

ListMFADevicesErrorKind

ListOpenIDConnectProviderTagsErrorKind

ListOpenIDConnectProvidersErrorKind

ListPoliciesErrorKind

ListPoliciesGrantingServiceAccessErrorKind

ListPolicyTagsErrorKind

ListPolicyVersionsErrorKind

ListRolePoliciesErrorKind

ListRoleTagsErrorKind

ListRolesErrorKind

ListSAMLProviderTagsErrorKind

ListSAMLProvidersErrorKind

ListSSHPublicKeysErrorKind

ListServerCertificateTagsErrorKind

ListServerCertificatesErrorKind

ListServiceSpecificCredentialsErrorKind

ListSigningCertificatesErrorKind

ListUserPoliciesErrorKind

ListUserTagsErrorKind

ListUsersErrorKind

ListVirtualMFADevicesErrorKind

PutGroupPolicyErrorKind

PutRolePermissionsBoundaryErrorKind

PutRolePolicyErrorKind

PutUserPermissionsBoundaryErrorKind

PutUserPolicyErrorKind

RemoveClientIDFromOpenIDConnectProviderErrorKind

RemoveRoleFromInstanceProfileErrorKind

RemoveUserFromGroupErrorKind

ResetServiceSpecificCredentialErrorKind

ResyncMFADeviceErrorKind

SetDefaultPolicyVersionErrorKind

SetSecurityTokenServicePreferencesErrorKind

SimulateCustomPolicyErrorKind

SimulatePrincipalPolicyErrorKind

TagInstanceProfileErrorKind

TagMFADeviceErrorKind

TagOpenIDConnectProviderErrorKind

TagPolicyErrorKind

TagRoleErrorKind

TagSAMLProviderErrorKind

TagServerCertificateErrorKind

TagUserErrorKind

UntagInstanceProfileErrorKind

UntagMFADeviceErrorKind

UntagOpenIDConnectProviderErrorKind

UntagPolicyErrorKind

UntagRoleErrorKind

UntagSAMLProviderErrorKind

UntagServerCertificateErrorKind

UntagUserErrorKind

UpdateAccessKeyErrorKind

UpdateAccountPasswordPolicyErrorKind

UpdateAssumeRolePolicyErrorKind

UpdateGroupErrorKind

UpdateLoginProfileErrorKind

UpdateOpenIDConnectProviderThumbprintErrorKind

UpdateRoleDescriptionErrorKind

UpdateRoleErrorKind

UpdateSAMLProviderErrorKind

UpdateSSHPublicKeyErrorKind

UpdateServerCertificateErrorKind

UpdateServiceSpecificCredentialErrorKind

UpdateSigningCertificateErrorKind

UpdateUserErrorKind

UploadSSHPublicKeyErrorKind

UploadServerCertificateErrorKind

UploadSigningCertificateErrorKind