Struct aws_sdk_iam::input::simulate_custom_policy_input::Builder

#[non_exhaustive]
pub struct Builder { /* fields omitted */ }

A builder for SimulateCustomPolicyInput

Implementations

impl Builder

pub fn policy_input_list(self, input: impl Into<String>) -> Self

pub fn set_policy_input_list(self, input: Option<Vec<String>>) -> Self

pub fn permissions_boundary_policy_input_list(
self,
input: impl Into<String>
) -> Self

pub fn set_permissions_boundary_policy_input_list(
self,
input: Option<Vec<String>>
) -> Self

pub fn action_names(self, input: impl Into<String>) -> Self

pub fn set_action_names(self, input: Option<Vec<String>>) -> Self

pub fn resource_arns(self, input: impl Into<String>) -> Self

pub fn set_resource_arns(self, input: Option<Vec<String>>) -> Self

pub fn resource_policy(self, input: impl Into<String>) -> Self

A resource-based policy to include in the simulation provided as a string. Each resource in the simulation is treated as if it had this policy attached. You can include only one resource-based policy in a simulation.

The regex pattern used to validate this parameter is a string of characters consisting of the following:

Any printable ASCII character ranging from the space character (\u0020) through the end of the ASCII character range

The printable characters in the Basic Latin and Latin-1 Supplement character set (through \u00FF)

The special characters tab (\u0009), line feed (\u000A), and carriage return (\u000D)

pub fn set_resource_policy(self, input: Option<String>) -> Self

pub fn resource_owner(self, input: impl Into<String>) -> Self

An ARN representing the AWS account ID that specifies the owner of any simulated resource that does not identify its owner in the resource ARN. Examples of resource ARNs include an S3 bucket or object. If ResourceOwner is specified, it is also used as the account owner of any ResourcePolicy included in the simulation. If the ResourceOwner parameter is not specified, then the owner of the resources and the resource policy defaults to the account of the identity provided in CallerArn. This parameter is required only if you specify a resource-based policy and account that owns the resource is different from the account that owns the simulated calling user CallerArn.

The ARN for an account uses the following syntax: arn:aws:iam::AWS-account-ID:root. For example, to represent the account with the 112233445566 ID, use the following ARN: arn:aws:iam::112233445566-ID:root.

pub fn set_resource_owner(self, input: Option<String>) -> Self

pub fn caller_arn(self, input: impl Into<String>) -> Self

The ARN of the IAM user that you want to use as the simulated caller of the API operations. CallerArn is required if you include a ResourcePolicy so that the policy's Principal element has a value to use in evaluating the policy.

You can specify only the ARN of an IAM user. You cannot specify the ARN of an assumed role, federated user, or a service principal.

pub fn set_caller_arn(self, input: Option<String>) -> Self

pub fn context_entries(self, input: impl Into<ContextEntry>) -> Self

pub fn set_context_entries(self, input: Option<Vec<ContextEntry>>) -> Self

pub fn resource_handling_option(self, input: impl Into<String>) -> Self

Specifies the type of simulation to run. Different API operations that support resource-based policies require different combinations of resources. By specifying the type of simulation to run, you enable the policy simulator to enforce the presence of the required resources to ensure reliable simulation results. If your simulation does not match one of the following scenarios, then you can omit this parameter. The following list shows each of the supported scenario values and the resources that you must define to run the simulation.

Each of the EC2 scenarios requires that you specify instance, image, and security-group resources. If your scenario includes an EBS volume, then you must specify that volume as a resource. If the EC2 scenario includes VPC, then you must supply the network-interface resource. If it includes an IP subnet, then you must specify the subnet resource. For more information on the EC2 scenario options, see Supported platforms in the Amazon EC2 User Guide.

EC2-Classic-InstanceStore

instance, image, security-group

EC2-Classic-EBS

instance, image, security-group, volume

EC2-VPC-InstanceStore

instance, image, security-group, network-interface

EC2-VPC-InstanceStore-Subnet

instance, image, security-group, network-interface, subnet

EC2-VPC-EBS

instance, image, security-group, network-interface, volume

EC2-VPC-EBS-Subnet

instance, image, security-group, network-interface, subnet, volume

pub fn set_resource_handling_option(self, input: Option<String>) -> Self

pub fn max_items(self, input: i32) -> Self

Use this only when paginating results to indicate the maximum number of items you want in the response. If additional items exist beyond the maximum you specify, the IsTruncated response element is true.

If you do not include this parameter, the number of items defaults to 100. Note that IAM might return fewer results, even when there are more results available. In that case, the IsTruncated response element returns true, and Marker contains a value to include in the subsequent call that tells the service where to continue from.

pub fn set_max_items(self, input: Option<i32>) -> Self

pub fn marker(self, input: impl Into<String>) -> Self

Use this parameter only when paginating results and only after you receive a response indicating that the results are truncated. Set it to the value of the Marker element in the response that you received to indicate where the next call should start.

pub fn set_marker(self, input: Option<String>) -> Self

pub fn build(self) -> Result<SimulateCustomPolicyInput, BuildError>

Consumes the builder and constructs a SimulateCustomPolicyInput