Module aws_sdk_iam::model

Modules

access_detail

See AccessDetail

access_key

See AccessKey

access_key_last_used

See AccessKeyLastUsed

access_key_metadata

See AccessKeyMetadata

attached_permissions_boundary

See AttachedPermissionsBoundary

attached_policy

See AttachedPolicy

context_entry

See ContextEntry

deletion_task_failure_reason_type

See DeletionTaskFailureReasonType

entity_details

See EntityDetails

entity_info

See EntityInfo

error_details

See ErrorDetails

evaluation_result

See EvaluationResult

group

See Group

group_detail

See GroupDetail

instance_profile

See InstanceProfile

list_policies_granting_service_access_entry

See ListPoliciesGrantingServiceAccessEntry

login_profile

See LoginProfile

managed_policy_detail

See ManagedPolicyDetail

mfa_device

See MfaDevice

open_id_connect_provider_list_entry

See OpenIdConnectProviderListEntry

organizations_decision_detail

See OrganizationsDecisionDetail

password_policy

See PasswordPolicy

permissions_boundary_decision_detail

See PermissionsBoundaryDecisionDetail

policy

See Policy

policy_detail

See PolicyDetail

policy_granting_service_access

See PolicyGrantingServiceAccess

policy_group

See PolicyGroup

policy_role

See PolicyRole

policy_user

See PolicyUser

policy_version

See PolicyVersion

position

See Position

resource_specific_result

See ResourceSpecificResult

role

See Role

role_detail

See RoleDetail

role_last_used

See RoleLastUsed

role_usage_type

See RoleUsageType

saml_provider_list_entry

See SamlProviderListEntry

server_certificate

See ServerCertificate

server_certificate_metadata

See ServerCertificateMetadata

service_last_accessed

See ServiceLastAccessed

service_specific_credential

See ServiceSpecificCredential

service_specific_credential_metadata

See ServiceSpecificCredentialMetadata

signing_certificate

See SigningCertificate

ssh_public_key

See SshPublicKey

ssh_public_key_metadata

See SshPublicKeyMetadata

statement

See Statement

tag

See Tag

tracked_action_last_accessed

See TrackedActionLastAccessed

user

See User

user_detail

See UserDetail

virtual_mfa_device

See VirtualMfaDevice

Structs

AccessDetail

An object that contains details about when a principal in the reported AWS Organizations entity

AccessKey

Contains information about an AWS access key.

AccessKeyLastUsed

Contains information about the last time an AWS access key was used since IAM began

AccessKeyMetadata

Contains information about an AWS access key, without its secret key.

AttachedPermissionsBoundary

Contains information about an attached permissions boundary.

AttachedPolicy

Contains information about an attached policy.

ContextEntry

Contains information about a condition context key. It includes the name of the key and

DeletionTaskFailureReasonType

The reason that the service-linked role deletion failed.

EntityDetails

An object that contains details about when the IAM entities (users or roles) were last

EntityInfo

Contains details about the specified entity (user or role).

ErrorDetails

Contains information about the reason that the operation failed.

EvaluationResult

Contains the results of a simulation.

Group

Contains information about an IAM group entity.

GroupDetail

Contains information about an IAM group, including all of the group's policies.

InstanceProfile

Contains information about an instance profile.

ListPoliciesGrantingServiceAccessEntry

Contains details about the permissions policies that are attached to the specified

LoginProfile

Contains the user name and password create date for a user.

ManagedPolicyDetail

Contains information about a managed policy, including the policy's ARN, versions, and

MfaDevice

Contains information about an MFA device.

OpenIdConnectProviderListEntry

Contains the Amazon Resource Name (ARN) for an IAM OpenID Connect provider.

OrganizationsDecisionDetail

Contains information about the effect that Organizations has on a policy simulation.

PasswordPolicy

Contains information about the account password policy.

PermissionsBoundaryDecisionDetail

Contains information about the effect that a permissions boundary has on a policy

Policy

Contains information about a managed policy.

PolicyDetail

Contains information about an IAM policy, including the policy document.

PolicyGrantingServiceAccess

Contains details about the permissions policies that are attached to the specified

PolicyGroup

Contains information about a group that a managed policy is attached to.

PolicyRole

Contains information about a role that a managed policy is attached to.

PolicyUser

Contains information about a user that a managed policy is attached to.

PolicyVersion

Contains information about a version of a managed policy.

Position

Contains the row and column of a location of a Statement element in a

ResourceSpecificResult

Contains the result of the simulation of a single API operation call on a single

Role

Contains information about an IAM role. This structure is returned as a response

RoleDetail

Contains information about an IAM role, including all of the role's policies.

RoleLastUsed

Contains information about the last time that an IAM role was used. This includes the

RoleUsageType

An object that contains details about how a service-linked role is used, if that

SamlProviderListEntry

Contains the list of SAML providers for this account.

ServerCertificate

Contains information about a server certificate.

ServerCertificateMetadata

Contains information about a server certificate without its certificate body,

ServiceLastAccessed

Contains details about the most recent attempt to access the service.

ServiceSpecificCredential

Contains the details of a service-specific credential.

ServiceSpecificCredentialMetadata

Contains additional details about a service-specific credential.

SigningCertificate

Contains information about an X.509 signing certificate.

SshPublicKey

Contains information about an SSH public key.

SshPublicKeyMetadata

Contains information about an SSH public key, without the key's body or

Statement

Contains a reference to a Statement element in a policy document that

Tag

A structure that represents user-provided metadata that can be associated with an IAM

TrackedActionLastAccessed

Contains details about the most recent attempt to access an action within the service.

User

Contains information about an IAM user entity.

UserDetail

Contains information about an IAM user, including all the user's policies and all the

VirtualMfaDevice

Contains information about a virtual MFA device.

Enums

AccessAdvisorUsageGranularityType

AssignmentStatusType

ContextKeyTypeEnum

DeletionTaskStatusType

EncodingType

EntityType

GlobalEndpointTokenVersion

JobStatusType

PermissionsBoundaryAttachmentType

PolicyEvaluationDecisionType

PolicyOwnerEntityType

PolicyScopeType

PolicySourceType

PolicyType

PolicyUsageType

The policy usage type that indicates whether the policy is used as a permissions policy

ReportFormatType

ReportStateType

SortKeyType

StatusType

SummaryKeyType