Struct aws_sdk_iam::model::ResourceSpecificResult

#[non_exhaustive]
pub struct ResourceSpecificResult {
    pub eval_resource_name: Option<String>,
    pub eval_resource_decision: Option<PolicyEvaluationDecisionType>,
    pub matched_statements: Option<Vec<Statement>>,
    pub missing_context_values: Option<Vec<String>>,
    pub eval_decision_details: Option<HashMap<String, PolicyEvaluationDecisionType>>,
    pub permissions_boundary_decision_detail: Option<PermissionsBoundaryDecisionDetail>,
}

Contains the result of the simulation of a single API operation call on a single resource.

This data type is used by a member of the EvaluationResult data type.

Fields (Non-exhaustive)

Non-exhaustive structs could have additional fields added in future. Therefore, non-exhaustive structs cannot be constructed in external crates using the traditional Struct { .. } syntax; cannot be matched against without a wildcard ..; and struct update syntax will not work.

eval_resource_name: Option<String>

The name of the simulated resource, in Amazon Resource Name (ARN) format.

eval_resource_decision: Option<PolicyEvaluationDecisionType>

The result of the simulation of the simulated API operation on the resource specified in EvalResourceName.

matched_statements: Option<Vec<Statement>>

A list of the statements in the input policies that determine the result for this part of the simulation. Remember that even if multiple statements allow the operation on the resource, if any statement denies that operation, then the explicit deny overrides any allow. In addition, the deny statement is the only entry included in the result.

missing_context_values: Option<Vec<String>>

A list of context keys that are required by the included input policies but that were not provided by one of the input parameters. This list is used when a list of ARNs is included in the ResourceArns parameter instead of "*". If you do not specify individual resources, by setting ResourceArns to "*" or by not including the ResourceArns parameter, then any missing context values are instead included under the EvaluationResults section. To discover the context keys used by a set of policies, you can call GetContextKeysForCustomPolicy or GetContextKeysForPrincipalPolicy.

eval_decision_details: Option<HashMap<String, PolicyEvaluationDecisionType>>

Additional details about the results of the evaluation decision on a single resource. This parameter is returned only for cross-account simulations. This parameter explains how each policy type contributes to the resource-specific evaluation decision.

permissions_boundary_decision_detail: Option<PermissionsBoundaryDecisionDetail>

Contains information about the effect that a permissions boundary has on a policy simulation when that boundary is applied to an IAM entity.

Implementations

impl ResourceSpecificResult

pub fn builder() -> Builder

Creates a new builder-style object to manufacture ResourceSpecificResult