Module aws_sdk_iam::operation

Structs

AddClientIDToOpenIDConnectProvider

Adds a new client ID (also known as audience) to the list of client IDs already

AddRoleToInstanceProfile

Adds the specified IAM role to the specified instance profile. An instance profile

AddUserToGroup

Adds the specified user to the specified group.

AttachGroupPolicy

Attaches the specified managed policy to the specified IAM group.

AttachRolePolicy

Attaches the specified managed policy to the specified IAM role. When you attach a

AttachUserPolicy

Attaches the specified managed policy to the specified user.

ChangePassword

Changes the password of the IAM user who is calling this operation. This operation

CreateAccessKey

Creates a new AWS secret access key and corresponding AWS access key ID for the

CreateAccountAlias

Creates an alias for your AWS account. For information about using an AWS account

CreateGroup

Creates a new group.

CreateInstanceProfile

Creates a new instance profile. For information about instance profiles, see Using

CreateLoginProfile

Creates a password for the specified IAM user. A password allows an IAM user to

CreateOpenIDConnectProvider

Creates an IAM entity to describe an identity provider (IdP) that supports OpenID Connect (OIDC).

CreatePolicy

Creates a new managed policy for your AWS account.

CreatePolicyVersion

Creates a new version of the specified managed policy. To update a managed policy, you

CreateRole

Creates a new role for your AWS account. For more information about roles, see

CreateSAMLProvider

Creates an IAM resource that describes an identity provider (IdP) that supports SAML

CreateServiceLinkedRole

Creates an IAM role that is linked to a specific AWS service. The service controls

CreateServiceSpecificCredential

Generates a set of credentials consisting of a user name and password that can be used

CreateUser

Creates a new IAM user for your AWS account.

CreateVirtualMFADevice

Creates a new virtual MFA device for the AWS account. After creating the virtual

DeactivateMFADevice

Deactivates the specified MFA device and removes it from association with the user

DeleteAccessKey

Deletes the access key pair associated with the specified IAM user.

DeleteAccountAlias

Deletes the specified AWS account alias. For information about using an AWS

DeleteAccountPasswordPolicy

Deletes the password policy for the AWS account. There are no parameters.

DeleteGroup

Deletes the specified IAM group. The group must not contain any users or have any

DeleteGroupPolicy

Deletes the specified inline policy that is embedded in the specified IAM

DeleteInstanceProfile

Deletes the specified instance profile. The instance profile must not have an

DeleteLoginProfile

Deletes the password for the specified IAM user, which terminates the user's ability

DeleteOpenIDConnectProvider

Deletes an OpenID Connect identity provider (IdP) resource object in IAM.

DeletePolicy

Deletes the specified managed policy.

DeletePolicyVersion

Deletes the specified version from the specified managed policy.

DeleteRole

Deletes the specified role. The role must not have any policies attached. For more

DeleteRolePermissionsBoundary

Deletes the permissions boundary for the specified IAM role.

DeleteRolePolicy

Deletes the specified inline policy that is embedded in the specified IAM

DeleteSAMLProvider

Deletes a SAML provider resource in IAM.

DeleteSSHPublicKey

Deletes the specified SSH public key.

DeleteServerCertificate

Deletes the specified server certificate.

DeleteServiceLinkedRole

Submits a service-linked role deletion request and returns a

DeleteServiceSpecificCredential

Deletes the specified service-specific credential.

DeleteSigningCertificate

Deletes a signing certificate associated with the specified IAM user.

DeleteUser

Deletes the specified IAM user. Unlike the AWS Management Console, when you delete a user

DeleteUserPermissionsBoundary

Deletes the permissions boundary for the specified IAM user.

DeleteUserPolicy

Deletes the specified inline policy that is embedded in the specified IAM

DeleteVirtualMFADevice

Deletes a virtual MFA device.

DetachGroupPolicy

Removes the specified managed policy from the specified IAM group.

DetachRolePolicy

Removes the specified managed policy from the specified role.

DetachUserPolicy

Removes the specified managed policy from the specified user.

EnableMFADevice

Enables the specified MFA device and associates it with the specified IAM user. When

GenerateCredentialReport

Generates a credential report for the AWS account. For more information about the

GenerateOrganizationsAccessReport

Generates a report for service last accessed data for AWS Organizations. You can generate a report

GenerateServiceLastAccessedDetails

Generates a report that includes details about when an IAM resource (user, group,

GetAccessKeyLastUsed

Retrieves information about when the specified access key was last used. The

GetAccountAuthorizationDetails

Retrieves information about all IAM users, groups, roles, and policies in your AWS

GetAccountPasswordPolicy

Retrieves the password policy for the AWS account. This tells you the complexity

GetAccountSummary

Retrieves information about IAM entity usage and IAM quotas in the AWS

GetContextKeysForCustomPolicy

Gets a list of all of the context keys referenced in the input policies. The policies

GetContextKeysForPrincipalPolicy

Gets a list of all of the context keys referenced in all the IAM policies that are

GetCredentialReport

Retrieves a credential report for the AWS account. For more information about the

GetGroup

Returns a list of IAM users that are in the specified IAM group. You can paginate

GetGroupPolicy

Retrieves the specified inline policy document that is embedded in the specified IAM

GetInstanceProfile

Retrieves information about the specified instance profile, including the instance

GetLoginProfile

Retrieves the user name and password creation date for the specified IAM user. If

GetOpenIDConnectProvider

Returns information about the specified OpenID Connect (OIDC) provider resource object

GetOrganizationsAccessReport

Retrieves the service last accessed data report for AWS Organizations that was previously generated

GetPolicy

Retrieves information about the specified managed policy, including the policy's

GetPolicyVersion

Retrieves information about the specified version of the specified managed policy,

GetRole

Retrieves information about the specified role, including the role's path, GUID, ARN,

GetRolePolicy

Retrieves the specified inline policy document that is embedded with the specified

GetSAMLProvider

Returns the SAML provider metadocument that was uploaded when the IAM SAML provider

GetSSHPublicKey

Retrieves the specified SSH public key, including metadata about the key.

GetServerCertificate

Retrieves information about the specified server certificate stored in IAM.

GetServiceLastAccessedDetails

Retrieves a service last accessed report that was created using the

GetServiceLastAccessedDetailsWithEntities

After you generate a group or policy report using the

GetServiceLinkedRoleDeletionStatus

Retrieves the status of your service-linked role deletion. After you use DeleteServiceLinkedRole to submit a service-linked role for deletion,

GetUser

Retrieves information about the specified IAM user, including the user's creation

GetUserPolicy

Retrieves the specified inline policy document that is embedded in the specified IAM

ListAccessKeys

Returns information about the access key IDs associated with the specified IAM user.

ListAccountAliases

Lists the account alias associated with the AWS account (Note: you can have only

ListAttachedGroupPolicies

Lists all managed policies that are attached to the specified IAM group.

ListAttachedRolePolicies

Lists all managed policies that are attached to the specified IAM role.

ListAttachedUserPolicies

Lists all managed policies that are attached to the specified IAM user.

ListEntitiesForPolicy

Lists all IAM users, groups, and roles that the specified managed policy is attached

ListGroupPolicies

Lists the names of the inline policies that are embedded in the specified IAM

ListGroups

Lists the IAM groups that have the specified path prefix.

ListGroupsForUser

Lists the IAM groups that the specified IAM user belongs to.

ListInstanceProfileTags

Lists the tags that are attached to the specified IAM instance profile. The returned list of tags is sorted by tag key.

ListInstanceProfiles

Lists the instance profiles that have the specified path prefix. If there are none,

ListInstanceProfilesForRole

Lists the instance profiles that have the specified associated IAM role. If there

ListMFADeviceTags

Lists the tags that are attached to the specified IAM virtual multi-factor authentication (MFA) device. The returned list of tags is

ListMFADevices

Lists the MFA devices for an IAM user. If the request includes a IAM user name,

ListOpenIDConnectProviderTags

Lists the tags that are attached to the specified OpenID Connect (OIDC)-compatible

ListOpenIDConnectProviders

Lists information about the IAM OpenID Connect (OIDC) provider resource objects

ListPolicies

Lists all the managed policies that are available in your AWS account, including

ListPoliciesGrantingServiceAccess

Retrieves a list of policies that the IAM identity (user, group, or role) can use to

ListPolicyTags

Lists the tags that are attached to the specified IAM customer managed policy.

ListPolicyVersions

Lists information about the versions of the specified managed policy, including the

ListRolePolicies

Lists the names of the inline policies that are embedded in the specified IAM

ListRoleTags

Lists the tags that are attached to the specified role. The returned list of tags is

ListRoles

Lists the IAM roles that have the specified path prefix. If there are none, the

ListSAMLProviderTags

Lists the tags that are attached to the specified Security Assertion Markup Language

ListSAMLProviders

Lists the SAML provider resource objects defined in IAM in the account.

ListSSHPublicKeys

Returns information about the SSH public keys associated with the specified IAM

ListServerCertificateTags

Lists the tags that are attached to the specified IAM server certificate. The

ListServerCertificates

Lists the server certificates stored in IAM that have the specified path prefix. If

ListServiceSpecificCredentials

Returns information about the service-specific credentials associated with the

ListSigningCertificates

Returns information about the signing certificates associated with the specified IAM

ListUserPolicies

Lists the names of the inline policies embedded in the specified IAM user.

ListUserTags

Lists the tags that are attached to the specified IAM user. The returned list of tags is sorted by tag key. For more information about tagging, see Tagging IAM resources in the

ListUsers

Lists the IAM users that have the specified path prefix. If no path prefix is

ListVirtualMFADevices

Lists the virtual MFA devices defined in the AWS account by assignment status. If

PutGroupPolicy

Adds or updates an inline policy document that is embedded in the specified IAM

PutRolePermissionsBoundary

Adds or updates the policy that is specified as the IAM role's permissions boundary.

PutRolePolicy

Adds or updates an inline policy document that is embedded in the specified IAM

PutUserPermissionsBoundary

Adds or updates the policy that is specified as the IAM user's permissions boundary.

PutUserPolicy

Adds or updates an inline policy document that is embedded in the specified IAM

RemoveClientIDFromOpenIDConnectProvider

Removes the specified client ID (also known as audience) from the list of client IDs

RemoveRoleFromInstanceProfile

Removes the specified IAM role from the specified EC2 instance profile.

RemoveUserFromGroup

Removes the specified user from the specified group.

ResetServiceSpecificCredential

Resets the password for a service-specific credential. The new password is AWS

ResyncMFADevice

Synchronizes the specified MFA device with its IAM resource object on the AWS

SetDefaultPolicyVersion

Sets the specified version of the specified policy as the policy's default (operative)

SetSecurityTokenServicePreferences

Sets the specified version of the global endpoint token as the token version used for

SimulateCustomPolicy

Simulate how a set of IAM policies and optionally a resource-based policy works with

SimulatePrincipalPolicy

Simulate how a set of IAM policies attached to an IAM entity works with a list of

TagInstanceProfile

Adds one or more tags to an IAM instance profile. If a tag with the same key name

TagMFADevice

Adds one or more tags to an IAM virtual multi-factor authentication (MFA) device. If

TagOpenIDConnectProvider

Adds one or more tags to an OpenID Connect (OIDC)-compatible identity provider. For

TagPolicy

Adds one or more tags to an IAM customer managed policy. If a tag with the same key

TagRole

Adds one or more tags to an IAM role. The role can be a regular role or a

TagSAMLProvider

Adds one or more tags to a Security Assertion Markup Language (SAML) identity provider.

TagServerCertificate

Adds one or more tags to an IAM server certificate. If a tag with the same key name

TagUser

Adds one or more tags to an IAM user. If a tag with the same key name already exists,

UntagInstanceProfile

Removes the specified tags from the IAM instance profile. For more information about tagging, see Tagging IAM resources in the

UntagMFADevice

Removes the specified tags from the IAM virtual multi-factor authentication (MFA)

UntagOpenIDConnectProvider

Removes the specified tags from the specified OpenID Connect (OIDC)-compatible identity

UntagPolicy

Removes the specified tags from the customer managed policy. For more information about tagging, see Tagging IAM resources in the

UntagRole

Removes the specified tags from the role. For more information about tagging, see Tagging IAM resources in the

UntagSAMLProvider

Removes the specified tags from the specified Security Assertion Markup Language (SAML)

UntagServerCertificate

Removes the specified tags from the IAM server certificate.

UntagUser

Removes the specified tags from the user. For more information about tagging, see Tagging IAM resources in the

UpdateAccessKey

Changes the status of the specified access key from Active to Inactive, or vice versa.

UpdateAccountPasswordPolicy

Updates the password policy settings for the AWS account.

UpdateAssumeRolePolicy

Updates the policy that grants an IAM entity permission to assume a role. This is

UpdateGroup

Updates the name and/or the path of the specified IAM group.

UpdateLoginProfile

Changes the password for the specified IAM user. You can use the AWS CLI, the

UpdateOpenIDConnectProviderThumbprint

Replaces the existing list of server certificate thumbprints associated with an OpenID

UpdateRole

Updates the description or maximum session duration setting of a role.

UpdateRoleDescription

Use UpdateRole instead.

UpdateSAMLProvider

Updates the metadata document for an existing SAML provider resource object.

UpdateSSHPublicKey

Sets the status of an IAM user's SSH public key to active or inactive. SSH public

UpdateServerCertificate

Updates the name and/or the path of the specified server certificate stored in

UpdateServiceSpecificCredential

Sets the status of a service-specific credential to Active or

UpdateSigningCertificate

Changes the status of the specified user signing certificate from active to disabled,

UpdateUser

Updates the name and/or the path of the specified IAM user.

UploadSSHPublicKey

Uploads an SSH public key and associates it with the specified IAM user.

UploadServerCertificate

Uploads a server certificate entity for the AWS account. The server certificate

UploadSigningCertificate

Uploads an X.509 signing certificate and associates it with the specified IAM user.